Refactor visit(JavadocComment) to be a real visitor method
This uses the Visitor pattern to change behavior depending on dynamic type, rather than an instanceof check.

The method now @Overrides the visitor method from the superclass, removing the need for the comment.

The getCommentTypeJavadoc instanceof check is removed as it is no longer needed.

Since CommentHelperMethods now no longer depends on the AnalysisVisitor, its methods can be static, so this makes it so.

Behavior is unchanged.