Add basic Maven plugin and usage example
The plugin can be bound to the process-sources Maven phase in a project and will then generate a livingdocumentation.json file in the target directory.

Any JARs that are part of the compilation classpath are added to the parser to resolve type information.

The current plugin is not really configurable; it always does the above thing.