Add sketch of a Kotlin analyzer
The Kotlin analyzer is based on the embeddable version of the actual Kotlin compiler. It is not yet strong enough to do type inference (which is harder/more powerful in Kotlin than it is in Java) but it does parse a bunch of things and generates JSON.

Instead of a data model as Kotlin classes, it calls the Jackson library to generate JSON directly. This is a bit more brittle.

This is undocumented, untested, and incomplete.