Add checkstyle Maven plugin
This uses Google style as a base, and fails the build if something is wrong.

Some of these warnings were actually failing, so this PR fixes those (mostly imports and test class names).

The style guide (in 7.3.1) gives an exception that you don't have to put a Javadoc comment on "simple, obvious methods like getFoo" - of course, the machine can't really discern what is simple or obvious, so this configuration disables that check; it needs to be checked during code review instead.