Add initial Gradle plugin
The Gradle plugin follows the behavior of the Maven plugin: if it is applied, it automatically generates a livingdocumentation.json file for the Java source that's in the project every time the project is compiled.

In addition, the task can be run manually with `./gradlew generateLivingDocumentation` or `./gradlew gLD` for short, or through IDE integration.

As a bonus, since Gradle tracks input and output files, this is a no-op if no files have changed since the last time the JSON was generated.

The plugin assumes that the 'java' plugin is also applied to the project.